+++ packages/client/src/admin/surveys/collectText.ts
@@ -1,0 +1,21 @@
+/**
+ * Extracts text from the given ProseMirror document up to the character limit
+ * @param body ProseMirror document json
+ * @param charLimit Character limit
+ * @returns string
+ */
+export function collectText(body: any, charLimit = 32) {
+  let text = "";
+  if (body.text) {
+    text += body.text + " ";
+  }
+  if (body.content && body.content.length) {
+    for (const node of body.content) {
+      if (text.length > charLimit) {
+        return text;
+      }
+      text += collectText(node);
+    }
+  }
+  return text;
+}
